requirements.txt
pystray==0.17.0
win10toast==0.9
pyinstaller==5.6.2
Pillow==8.3.2
